Create NUMBER GAME
Generate a random number within a specified range, such as 1 to 100.
Prompt the user to enter their guess for the generated number.
Compare the user
'
s guess with the generated number and provide

feedback on whether the guess is correct, too high, or too low.
Repeat steps 2 and 3 until the user guesses the correct number. You can
incorporate additional details as follows:
Limit the number of attempts the user has to guess the number.
Add the option for multiple rounds, allowing the user to play again.
Display the user
'
s score, which can be based on the number of attempts

taken or rounds won.